chore(evm): shadow hardhat in the evm repo
Hardhat's presence (specifically the npm solady package and the default
    config file) breaks the nomic labs LSP when using foundry, which adds
some churn to working on contracts. I attempted to completely remove
hardhat from the evm package but I'd have to refactor significant portions
of the the sdk and build pipelines to get it all working again. I'm not
bothered that hardhat is used in the build pipeline, so I just shadowed
the config file behind a non-default name and removed some unused
scripts.