chore(evm): improve turbo build caching
turbo integrates well with the vcs and npm package.json files to
determine caching behavior, but we had a nested gitignore file that
turbo couldn't observe. The purpose of these changes is to improve the
turbo cache for the evm package, which is a build dependency for all
other packages. This should significantly speed up workflows with more
cache catches, which will allow for faster stack-based merges and local
testing