fix(evm): enable multiple claims by an address for ERC1155Incentives
Claimants can claim 1 ERC 1155 per transaction. Transactions are
validated offchain, but to ensure a unique signanature and claim
attempt, transaction hashes are passed in as part of the claim payload
and stored onchain. As a basic validation with `isClaimable`, the same
transaction hash cannot be claimed against, irrespective of the claim
target