test(evm): create test for signature malleability
We are not cutting of the upper half of the curve, so it's technically
feasible to submit a modified signature that will pass validation with
`ecrecover`. However, other measures and validations ensure that this
doesn't amount to a security risk. First, eip-712 ensures that the
signature cannot be replayed across chains. Second, we are not storing
signatures as keys anywhere and instead use the hash payloads, rendering
signature replay on the same chain impossible.

The _only_ risk involving signautures entails approving a boost claim
with multiple incentives before the claimant is eligible to claim all
incentives, since we sign off on incentive quantity instead of individual
incentives. However, this would be an issue regardless of whether or not
the signature was malleable.